Use u64 for Signatory voting power, and u128 for SignatorySet total voting power (to prevent overflow). Also added SignatorySet::two_thirds_voting_power().

+++ chain/chain/src/state_machine.rs
@@ -176,7 +176,7 @@ fn signatories_from_validators(validators: &BTreeMap<Vec<u8>, u64>) -> Result<Si
     let mut signatories = SignatorySet::new();
     for (key_bytes, voting_power) in validators.iter() {
         let key = bitcoin::PublicKey::from_slice(key_bytes.as_slice())?;
-        signatories.set(Signatory::new(key, *voting_power as u32));
+        signatories.set(Signatory::new(key, *voting_power));
     }
     Ok(signatories)
 }
